# docs/api.yaml
openapi: 3.1.0
info:
  title: TeeReserve Wireframe API v1
  version: 1.0.0
  description: API canónico con endpoints de cursos, disponibilidad, tarifas, pagos, reservas y recibos.
servers:
  - url: http://localhost:3002/api/v1
paths:
...
  /courses/{id}/availability:
    get:
      summary: Disponibilidad de tee times
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: Fecha YYYY-MM-DD
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AvailabilitySlot'
  /courses/{id}/rates:
    get:
      summary: Tarifas de un curso
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Rate'
...
components:
  schemas:
...
    AvailabilitySlot:
      type: object
      properties:
        date:
          type: string
        teeTime:
          type: string
        playersMin:
          type: number
        playersMax:
          type: number
        publicPriceUSD:
          type: number
        currency:
          type: string
    Rate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        publicPriceUSD:
          type: number
        currency:
          type: string
...